feat(argparse): add argument parser option for number of holes
Added another argument parser option for the number of holes and limited the range to be only
between 3..=6. This range was again based on the Wikipedia article for the Mastermind board game.
There were quite a few options in the table listing the number of holes but the generally range was
between three and six.