fix: bug with counting number of incorrect color code pegs
The number of color code pegs could go to zero sometimes and that condition was not being checked.
Due to the nature of the frequency being in terms of unsigned size, this cause a backflow. The
backflow would then cause the program to fail. This condition should prevent that from happening in
the future.